pyproject.toml
[project]
name = "notesrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.59.0",
    "google-genai-haystack>=3.11.0",
    "haystack-ai>=2.26.1",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langsmith[claude-agent-sdk]>=0.7.25",
    "pinecone-haystack>=6.1.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
